move queries string to new file
